Move skybox and envmap to framework/scene.
