+++ src/framework/shaders/envmap/irradiance_calculate_coeff.comp.glsl
@@ -11,8 +11,7 @@
 //
 // ----------------------------------------------------------------------------
 
-#include "../envmap_interop.h"
-
+#include <envmap/interop.h>
 #include <shared/maths.glsl>
 
 // ----------------------------------------------------------------------------
