+++ src/framework/shaders/envmap/irradiance_reduce_step.comp.glsl
@@ -4,7 +4,7 @@
 
 // ----------------------------------------------------------------------------
 
-#include "../envmap_interop.h"
+#include <envmap/interop.h>
 
 // ----------------------------------------------------------------------------
