+++ src/framework/shaders/envmap/irradiance_transfer_coeff.comp.glsl
@@ -4,8 +4,7 @@
 
 // ----------------------------------------------------------------------------
 
-#include "../envmap_interop.h"
-
+#include <envmap/interop.h>
 #include <shared/constants.glsl>
 
 // ----------------------------------------------------------------------------
