+++ src/framework/shaders/envmap/specular_convolution.comp.glsl
@@ -16,8 +16,7 @@
 //
 // ----------------------------------------------------------------------------
 
-#include "../envmap_interop.h"
-
+#include <envmap/interop.h>
 #include <shared/maths.glsl>
 
 // ----------------------------------------------------------------------------
