+++ src/framework/shaders/skybox/integrate_brdf.comp.glsl
@@ -15,8 +15,7 @@
 //
 // ----------------------------------------------------------------------------
 
-#include "../skybox_interop.h" //
-
+#include <skybox/interop.h>
 #include <shared/maths.glsl>        // for importance_sample_GGX
 #include <shared/lighting/pbr.glsl> // for gf_SmithGGX
