+++ src/framework/shaders/skybox/skybox.vert.glsl
@@ -4,7 +4,7 @@
 
 // ----------------------------------------------------------------------------
 
-#include "../skybox_interop.h"
+#include <skybox/interop.h>
 
 // ----------------------------------------------------------------------------
